♻️Refactor: game관련 api 정리
