+++ src/api/GameAPI/AddGameAPI.js
@@ -1,57 +1,6 @@
 import { GET_API } from '../CommonAPI';
-import tokenData from '../../assets/data/sport_users.json';
-
-// get TEAM IMAGE from team tokenData
-const getHomeImage = (team_name) => {
-  const team = tokenData.filter(
-    (item) => item.username.split(' ')[0] === team_name,
-  );
-  return team[0].image;
-};
-
-// get TEAM NAME from team tokenData
-const getTeamName = () => {
-  const team = tokenData.map((item) => item.username.split(' ')[0]);
-  team.unshift('전체');
-  return team;
-};
-
-// Conver Arr to Game Object
-const arrToGame = (gameArr) => {
-  const image = getHomeImage(gameArr[3]);
-  return {
-    'date': gameArr[0],
-    'day': gameArr[1],
-    'time': gameArr[2],
-    'home': gameArr[3],
-    'away': gameArr[4],
-    'stadium': gameArr[5],
-    'full_stadium': gameArr[6],
-    'en_city': gameArr[7],
-    'image': image,
-  };
-};
-
-// date / time 오름차순 정렬
-const sortGameByDate = (gameArr) => {
-  return gameArr.sort((a, b) => {
-    const infoA = a[0].split(',');
-    const infoB = b[0].split(',');
-    const dateA = new Date(infoA[0]);
-    const dateB = new Date(infoB[0]);
-    if (dateA > dateB) {
-      return 1;
-    } else if (dateA < dateB) {
-      return -2;
-    } else {
-      if (infoA[2] > infoB[2]) {
-        return 1;
-      } else {
-        return -2;
-      }
-    }
-  });
-};
+import { arrToGame } from '../../util/gameUtil';
+import { sortGameByDate } from '../../util/gameUtil';
 
 // get Game info with API
 // sort them in increasing order of date and time
@@ -79,24 +28,10 @@ const getGameInfo = async (token) => {
     }
   });
 
+  console.log(Array.from(gameMap));
   // date / time 오름차순 정렬
   const gameArr = sortGameByDate(Array.from(gameMap));
   return gameArr.map((item) => [arrToGame(item[0].split(',')), item[1]]);
 };
 
-// filter by type and team
-const filterGameInfo = (game, type, team) => {
-  if (team === '전체') {
-    return game;
-  }
-  return game.filter((item) => item[0].home === team || item[0].away === team);
-};
-
-export {
-  getGameInfo,
-  getHomeImage,
-  getTeamName,
-  filterGameInfo,
-  sortGameByDate,
-  arrToGame,
-};
+export { getGameInfo };
